requirements.txt
asgiref==3.6.0
boto3==1.26.135
botocore==1.29.135
Django==4.2.1
django-allow-cidr==0.6.0
django-cors-headers==4.0.0
django-environ==0.10.0
django-ses==3.4.1
django-storages==1.13.2
djangorestframework==3.14.0
jmespath==1.0.1
mysqlclient==2.1.1
packaging==23.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15